refactor: Hid DB processing in repository.
