+++ src/server/controllers/users.ts
@@ -1,10 +1,9 @@
 import express, { Request, Response } from 'express';
 import { InvalidCredentialsException } from '../../shared/exceptions/invalidCredentials';
 import { UnprocessableEntityException } from '../../shared/exceptions/unprocessableEntity';
-import { User } from '../../shared/types';
-import { getDatabase } from '../database/connection';
 import asyncHandler from '../middleware/asyncHandler';
 import permissionsHandler from '../middleware/permissionsHandler';
+import { UsersRepository } from '../repositories/users';
 import { oneWayHash } from '../utilities/oneWayHash';
 
 const app = express();
@@ -17,16 +16,9 @@ app.get(
       throw new InvalidCredentialsException('invalid_user_credentials');
     }
 
-    const database = getDatabase();
-    const users = await database
-      .select('u.*', {
-        roleId: 'r.id',
-        roleName: 'r.name',
-        roleDescription: 'r.description',
-        roleAdminAccess: 'r.admin_access',
-      })
-      .from('superfast_users AS u')
-      .join('superfast_roles AS r', 'r.id', 'u.role_id');
+    const repository = new UsersRepository();
+
+    const users = await repository.readWithRole();
 
     res.json({
       users: users.flatMap(({ password, ...user }) => payload(user)),
@@ -38,20 +30,10 @@ app.get(
   '/users/:id',
   permissionsHandler([{ collection: 'superfast_users', action: 'read' }]),
   asyncHandler(async (req: Request, res: Response) => {
-    const database = getDatabase();
-    const id = req.params.id;
+    const id = Number(req.params.id);
+    const repository = new UsersRepository();
 
-    const user = await database
-      .select('u.*', {
-        roleId: 'r.id',
-        roleName: 'r.name',
-        roleDescription: 'r.description',
-        roleAdminAccess: 'r.admin_access',
-      })
-      .from('superfast_users AS u')
-      .join('superfast_roles AS r', 'r.id', 'u.role_id')
-      .where('u.id', id)
-      .first();
+    const user = await repository.readOneWithRole({ id });
 
     if (!user) return res.status(400).end();
 
@@ -65,13 +47,13 @@ app.post(
   '/users',
   permissionsHandler([{ collection: 'superfast_users', action: 'create' }]),
   asyncHandler(async (req: Request, res: Response) => {
-    const database = getDatabase();
-    req.body.password = await oneWayHash(req.body.password);
+    const repository = new UsersRepository();
 
-    const users = await database<User>('superfast_users').insert(req.body, 'id');
+    req.body.password = await oneWayHash(req.body.password);
+    const user = await repository.create(req.body);
 
     res.json({
-      user: users[0],
+      user: user,
     });
   })
 );
@@ -80,14 +62,14 @@ app.patch(
   '/users/:id',
   permissionsHandler([{ collection: 'superfast_users', action: 'update' }]),
   asyncHandler(async (req: Request, res: Response) => {
-    const database = getDatabase();
     const id = Number(req.params.id);
+    const usersRepository = new UsersRepository();
 
     if (req.body.password) {
       req.body.password = await oneWayHash(req.body.password);
     }
 
-    await database('superfast_users').where('id', id).update(req.body);
+    await usersRepository.update(id, req.body);
 
     res.status(204).end();
   })
@@ -97,8 +79,8 @@ app.delete(
   '/users/:id',
   permissionsHandler([{ collection: 'superfast_users', action: 'delete' }]),
   asyncHandler(async (req: Request, res: Response) => {
-    const database = getDatabase();
     const id = Number(req.params.id);
+    const usersRepository = new UsersRepository();
 
     if (!req.userId) {
       throw new InvalidCredentialsException('invalid_user_credentials');
@@ -108,7 +90,7 @@ app.delete(
       throw new UnprocessableEntityException('can_not_delete_itself');
     }
 
-    await database('superfast_users').where('id', id).delete();
+    await usersRepository.delete(id);
 
     res.status(204).end();
   })
